anonymized gamebutton
- removed all references to parent structures from GameButton without altering functionality
- receives callables rather than objects with callables
