+++ tools/ui_widgets.py
@@ -1,10 +1,19 @@
 '''Used for easily initializing widgets, meant to reduce the repeated calls which specify the respective fields of similar widgets'''
 # pylint: disable=C0321, no-name-in-module, too-many-arguments
-from typing import List, Tuple, Optional, Union
+from enum import Enum
+from itertools import cycle
+from typing import List, Tuple, Optional, Union, Callable
 from PyQt5.QtWidgets import QPushButton, QLabel, QGroupBox, QGridLayout, QVBoxLayout, QToolButton, QLineEdit, QTextEdit, QRadioButton
 from PyQt5.QtGui import QValidator
 from PyQt5.QtCore import Qt
-from tools.stylesheets import BTN_STYLE_REF
+from tools.stylesheets import SELECT, get_btn_style, BTN_STYLE_REF
+
+class TURN(Enum):
+    '''Enum for tracking turn player'''
+    PL_1 = 1
+    PL_2 = 2
+
+TURN_CYCLER = cycle((TURN.PL_1, TURN.PL_2))
 
 class MenuButton(QPushButton):
     """A standardized button for the game"""
@@ -16,12 +25,36 @@ class MenuButton(QPushButton):
 
 class GameButton(QPushButton):
     """A standardized button for the game"""
-    def __init__(self, text: str = '', style: str = '', width: Optional[int] = None, height: Optional[int] = None, layout_position: Tuple[int,...] = (0,0)):
+    def __init__(self,
+                 text: str,
+                 size: int,
+                 layout_position: Tuple[int,int],
+                 parent_position: Tuple[int,int],
+                 get_turn_player: Callable[['GameButton'], TURN],
+                 registration: Callable[[Tuple[int,int],'GameButton'], None],
+                 ):
         super().__init__(text = text)
-        self.setStyleSheet(style)
-        if width: self.setFixedWidth(width)
-        if height: self.setFixedHeight(height)
+        self.setFixedSize(size, size)
         self.layout_position = layout_position
+        self.take_turn = get_turn_player
+        registration(parent_position, self)
+
+        self.setStyleSheet(get_btn_style(SELECT.DEFAULT_BTN))
+        self.is_claimed: bool = False
+
+    def claim(self):
+        '''a user has selected this square, gaining control of it'''
+        self.is_claimed = True
+        turn = self.take_turn(self)
+        ref = SELECT.PL_1 if turn is TURN.PL_1 else SELECT.PL_2
+        self.setStyleSheet(get_btn_style(ref))
+        self.setDisabled(True)
+
+    def resetButtonstate(self):
+        '''undo button was clicked, reset this square'''
+        self.is_claimed = False
+        self.setStyleSheet(get_btn_style(SELECT.DEFAULT_BTN))
+        self.setEnabled(True)
 
 class GameGroupBox(QGroupBox):
     """A standard groub box widget for the WAAM GUI"""
